test: Use cache folder set by common-tap.js
This sets the stage for checking to ensure that no root-owned
files end up being left in the cache folder after performing
any of the commands in any of the tests.

This also removes a lot of now-unnecessary setup and teardown
that various tests were doing, since the common script does it
all for them anyway.