!!! WIP --- XXX --- do not merge
Ok, found it, leaving this here as a note to myself and any other kind
soul who would like to assist in this.

The issue is that cacache.rm() will add a file to the _index_, even if
it's removing the content.  cacache.rm() is called when we make a
request that shouldn't be cached.  (To ensure that the content isn't
cached from some prior request, I guess?)

And, the uid/gid options **are not passed to cacache.rm**, which is a
big problem, because root-owned indexes are part of the issue that's
biting people so often.

Solution:

1. cacache.rm has to take an options object, pass to the index function
2. make-fetch-happen has to pass uid and gid to cacache.rm and
   cacache.get.info, since these can both write indexes.
3. npm-registry-fetch has to infer uid/gid from SUDO_* (as the default,
   at least), and pass that through to make-fetch-happen
4. All the libnpm* modules, they have to accept the uid/gid on their
   figgy puddings, and pass that through to cacache and
   npm-registry-fetch
5. All the lib/*.js files that call into those modules have to pull the
   uid/gid off of npmConfig in their figgy puddings

This commit does 1-3 in place (ie, not in the module's actual repos
where it needs to), and does 4 and 5 for access (but not all the other
commands).  This makes test/tap/access.js pass in sudo mode.

There are several others that need to be plumbed through.